Refactor: move class and package
