+++ src/main/java/me/kenux/travelog/domain/member/service/UserDetailsServiceImpl.java
@@ -1,4 +1,4 @@
-package me.kenux.travelog.global.security;
+package me.kenux.travelog.domain.member.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -6,7 +6,7 @@ import me.kenux.travelog.domain.member.entity.Member;
 import me.kenux.travelog.domain.member.repository.MemberRepository;
 import me.kenux.travelog.global.exception.CustomException;
 import me.kenux.travelog.global.exception.ErrorCode;
-import org.springframework.security.authentication.BadCredentialsException;
+import me.kenux.travelog.domain.member.service.dto.UserDetailsImpl;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
