+++ src/main/java/me/kenux/travelog/domain/member/service/dto/request/LogoutRequest.java
@@ -1,4 +1,4 @@
-package me.kenux.travelog.domain.member.dto.request;
+package me.kenux.travelog.domain.member.service.dto.request;
 
 import lombok.Data;
