+++ src/main/java/me/kenux/travelog/web/admin/MemberManageController.java
@@ -2,7 +2,7 @@ package me.kenux.travelog.web.admin;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import me.kenux.travelog.domain.member.dto.response.MemberInfoResponse;
+import me.kenux.travelog.domain.member.service.dto.response.MemberInfoResponse;
 import me.kenux.travelog.domain.member.repository.dto.MemberSearchCond;
 import me.kenux.travelog.domain.member.service.MemberService;
 import org.springframework.http.ResponseEntity;
