+++ src/main/java/me/kenux/travelog/web/api/member/MemberController.java
@@ -1,7 +1,7 @@
 package me.kenux.travelog.web.api.member;
 
 import lombok.RequiredArgsConstructor;
-import me.kenux.travelog.domain.member.dto.response.MemberInfoResponse;
+import me.kenux.travelog.domain.member.service.dto.response.MemberInfoResponse;
 import me.kenux.travelog.domain.member.service.MemberService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
